+++ src/pages/Detail/index.js
@@ -3,6 +3,6 @@ import { View } from 'react-native';
 
 export default function Detail() {
  return (
-   <View/>
+  
   );
 }
